:bug::scissors: Fix crash and moving some files
