+++ baseui/build.gradle
@@ -10,7 +10,7 @@ android {
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
-        versionName "1.0.0"
+        versionName "1.0.1"
     }
 
 }
@@ -18,8 +18,6 @@ android {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$rootProject.ext.kotlin_version"
-
-    implementation project(':common')
 
     // Support Libraries
     implementation "androidx.appcompat:appcompat:$rootProject.ext.androidXVersion"
