--- common/src/main/java/com/mytv/common/VersionUtils.kt
@@ -1,5 +1,0 @@
-package com.mytv.common
-
-import android.os.Build
-
-fun isMarshmallow() = Build.VERSION.SDK_INT >= Build.VERSION_CODES.M
