+++ settings.gradle
@@ -7,6 +7,5 @@ include ':home'
 include ':topseries'
 include ':data'
 include ':basetest'
-include ':common'
 include ':configuration'
 include ':images'
