+++ tools/global-dependencies.gradle
@@ -8,9 +8,9 @@ ext {
     gradleBuildToolsVersion = '3.4.0'
 
     // Internal
-    mytvBaseVersion = '1.0.0@aar'
+    mytvBaseVersion = '1.0.1@aar'
     mytvBaseTestVersion = '1.0.0@aar'
-    mytvBaseUiVersion = '1.0.0@aar'
+    mytvBaseUiVersion = '1.0.1@aar'
     mytvCommonVersion = '1.0.0@aar'
     mytvDataVersion = '1.0.0@aar'
